requirements.txt
Jinja2~=3.1.1
boto3~=1.21.36
ydb~=2.3.1
fire~=0.4.0
yandexcloud
bcrypt
PyJWT~=2.3.0